fix: Correctly Convert Text on all Cases
The input for `createMessage` in threads can either be a `string`
or a list of pre-determinated objects, one of which is a
`TextContentBlockParam`.

Currently, we are only handling the logic
for when the input is a `string`, which we correctly convert into
an output `Message`. However, the code breaks when the user inputs
a `TextContentBlockParam`, which is very useuful in situtations
such as sending text and an image in the same message.

This fix handles the case for when a `TextContentBlockParam`
is inputed, correctly converting it into a valid output `Messages`